Addressing more changes; Added new MishapEntityNotFound mishap documentation in Patchouli, among other things

+++ Common/src/main/java/at/petrak/hexcasting/api/casting/iota/DoubleIota.java
@@ -25,7 +25,7 @@ public class DoubleIota extends Iota {
     }
 
     public double getDouble() {
-        return HexUtils.fixNAN((Double) value);
+        return HexUtils.fixNAN(value);
     }
 
     @Override
